# crates/server/openapi.yaml
...
openapi: 3.0.3
info:
  title: Remote Harness Control Plane API
  version: 0.1.0
  description: |
    Control plane REST API. Human-readable API prose and behavior notes live in
    `docs/API_OVERVIEW.md`; this spec is the machine-readable contract for codegen and review.
servers:
  - url: /
    description: Relative to configured control plane base URL
...
paths:
...
  /api-keys/bootstrap:
    post:
      tags: [apiKeys]
      summary: Create the first API key when none exist (no authentication)
      operationId: bootstrapApiKey
      description: |
        Unauthenticated root-equivalent until a key exists. Returns 403 when `API_KEY`/`API_KEYS`
        is set or when any non-revoked key exists. See docs/API_OVERVIEW.md — Bootstrap safety.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: First API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreatedResponse"
        "403":
          description: Bootstrap disabled (keys already exist or env bootstrap key configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured or unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
  /api-keys:
    get:
      tags: [apiKeys]
      summary: List API keys
      operationId: listApiKeys
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated API keys (no secrets)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedApiKeySummaries"
        "400":
          description: Invalid cursor or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
    post:
      tags: [apiKeys]
      summary: Create API key
      operationId: createApiKey
      security:
        - bearerApiKey: []
        - headerApiKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created (plain key returned once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreatedResponse"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
  /api-keys/{id}:
    delete:
      tags: [apiKeys]
      summary: Revoke API key
      operationId: deleteApiKey
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Key revoked
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: "API key as Bearer token (`Authorization: Bearer <api_key>`)"
    headerApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key in `X-API-Key` header"
  schemas:
...
    StandardError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: object
              additionalProperties: true
...
    CreateApiKeyRequest:
      type: object
      properties:
        label:
          type: string
          nullable: true
    ApiKeyCreatedResponse:
      type: object
      required: [id, key, created_at]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          description: Plain secret; shown once at creation
        label:
          type: string
          nullable: true
        created_at:
          type: string
          description: ISO 8601 timestamp
    ApiKeySummary:
      type: object
      required: [id, created_at]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          nullable: true
        created_at:
          type: string
    PaginatedApiKeySummaries:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApiKeySummary"
        next_cursor:
          type: string
          nullable: true
...